build(migrations): add db migrations and instructions

+++ setup.sh
@@ -1,20 +1,31 @@
 #!/bin/sh
 
-EXPECTED_SIGNATURE="$(wget -q -O - https://composer.github.io/installer.sig)"
-php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-ACTUAL_SIGNATURE="$(php -r "echo hash_file('sha384', 'composer-setup.php');")"
+COMPOSER=bin/composer
 
-if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
-then
-    >&2 echo 'ERROR: Invalid installer signature'
-    rm composer-setup.php
-    exit 1
+if [ ! -f "$COMPOSER" ]; then
+  echo "* Installing Composer..."
+  EXPECTED_SIGNATURE="$(wget -q -O - http://composer.github.io/installer.sig)"
+  php -r "copy('http://getcomposer.org/installer', 'composer-setup.php');"
+  ACTUAL_SIGNATURE="$(php -r "echo hash_file('sha384', 'composer-setup.php');")"
+
+  if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
+  then
+      >&2 echo 'ERROR: Invalid installer signature'
+      rm composer-setup.php
+      exit 1
+  fi
+
+  php composer-setup.php --quiet
+  RESULT=$?
+  rm composer-setup.php
+  mv composer.phar bin/composer
 fi
 
-php composer-setup.php --quiet
-RESULT=$?
-rm composer-setup.php
-mv composer.phar bin/composer
+echo "* Installing packages..."
 bin/composer install
+echo "\n* Running database migrations..."
+vendor/bin/phinx migrate
+echo "\n* Setting file and folder permissions..."
 sh set_permissions.sh
+
 exit $RESULT
